fix: ConcatCol binding totalWidth 改用 size_t，消除 size_t 列宽累加进 uint32_t 的截断隐患

+++ xlite/csrc/_C.cpp
@@ -2140,7 +2140,7 @@ void ConcatCol(XRuntime &rt, std::vector<at::Tensor> &inputs, at::Tensor &out)
         return;
     }
     std::vector<XTensor> _inputs;
-    uint32_t totalWidth = 0;
+    size_t totalWidth = 0;
     for (auto &input : inputs) {
         XTensor x;
         InitXTensor(x, input);
